Major refactor
Create a monad for decimal arithmetic, and implement the numeric class
methods using it. Eliminate the context field from the Number type, only
using it within the new Arith monad. N.B. Trap handling is currently
unsatisfactory and will be revised.

Add numerous additional doctests.

Add instances for the remaining numeric type classes, and fix the Enum
instance. Note that the exponential and trigonometric functions are not
yet implemented.