Change return type of boolean operations
Per page 41 of the General Decimal Arithmetic Specification, operations
are allowed to return types other than decimal numbers or strings as
long as a conversion operation is provided.

Change the boolean operations to return Bool instead of Decimal, and add
a function fromBool to convert this to Decimal if desired.