Change compare to return Ordering
The compare and compareSignal operations now return Either (Decimal p r)
Ordering for convenience. Left values accommodate NaN results.

Update the Eq and Ord instances accordingly.

Improve the Ord compare method by resorting to a total ordering in the
case of NaN values.

Relax the constraints on the Ord instance, as well as the compare,
compareSignal, max, maxMagnitude, min, and minMagnitude operations not
to require Precision or Rounding on any type variable.