Revamp encoding types
Instead of newtypes, use generalized encoding format parameters that are
given a Precision instance so they can be used as any other Decimal
numbers.

Also move eMax and eMin to become Precision methods, and change the
Exponent type to Integer for the time being so we do not worry about
overflows.