Fix decodeFloat implementation
We had not scaled the m returned in (m, n) to be in the proper range
such that 10^(d-1) <= abs m < 10^d where d is the precision of the
number being decoded.

We do so now, and also update the doctests to check for this and a few
other properties.